+++ sdk-libs/transaction/tests/wallet_prop.rs
@@ -2,8 +2,10 @@ mod common;
 
 use std::collections::HashMap;
 
+use common::helpers::{
+    build_transfer, keypair_from_index, unique31, unique_nullifier, TransferSpec,
+};
 use common::InMemoryWallet;
-use common::{build_transfer, keypair_from_index, unique31, unique_nullifier, TransferSpec};
 use proptest::prelude::*;
 use proptest::test_runner::TestCaseError;
 use zolana_keypair::{ShieldedKeypair, SigningKey, ViewingKey};
